# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/ticker_events": {
            "post": {
                "description": "Inserts a ticker event segment for a given ticker and event.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker_events"
                ],
                "summary": "Add a new ticker event segment",
                "parameters": [
                    {
                        "description": "Ticker Event payload",
                        "name": "tickerEvent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PostTickerEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TickerEvent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ticker_events/{ticker}": {
            "get": {
                "description": "Returns all ticker event segments associated with a ticker symbol.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker_events"
                ],
                "summary": "List ticker events for a ticker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TickerEvent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/beta/{ticker}": {
            "get": {
                "description": "Retrieves the beta value and risk category of a given ticker symbol using Finnhub API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Get beta and risk classification for a stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock ticker symbol",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/comments": {
            "post": {
                "description": "Inserts a comment for the given ticker event segment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Add a new comment",
                "parameters": [
                    {
                        "description": "Comment payload",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PostCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Comment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/comments/{tickerEventID}": {
            "get": {
                "description": "Returns all comments associated with a ticker event ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "List comments for an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker Event ID",
                        "name": "tickerEventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Comment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/events/{event_id}": {
            "get": {
                "summary": "Gets events for a given ticker and timeframe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time in RFC3339 format",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End time in RFC3339 format",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of events to return",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "Username and password",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Log out a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "Returns matching ticker objects whose symbol or company name contains the query string.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search for tickers by symbol or company name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query string (part of ticker symbol or company name)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results grouped by type (e.g., stocks)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Ticker"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing or invalid query parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error (DB or API failure)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sign up a new user account",
                "parameters": [
                    {
                        "description": "Username, email and password",
                        "name": "signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/watchlist": {
            "get": {
                "summary": "Retrieve the tickers from current user's watchlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "token",
                        "in": "header"
                    }
                ],
                "responses": {}
            }
        },
        "/watchlist/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new ticker to current user's watchlist",
                "parameters": [
                    {
                        "description": "Ticker",
                        "name": "ticker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "token",
                        "in": "header"
                    }
                ],
                "responses": {}
            }
        },
        "/watchlist/remove": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new ticker to current user's watchlist",
                "parameters": [
                    {
                        "description": "Ticker",
                        "name": "ticker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "token",
                        "in": "header"
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "handler.PostCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "This drop makes sense after Q2 miss."
                },
                "ticker_event_id": {
                    "type": "string",
                    "example": "drop1"
                },
                "user_id": {
                    "type": "string",
                    "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
                }
            }
        },
        "handler.PostTickerEventRequest": {
            "type": "object",
            "properties": {
                "end_index": {
                    "type": "integer",
                    "example": 50
                },
                "event_id": {
                    "type": "string",
                    "example": "e7b8c6e5-3e44-4e2a-a6e7-123456789abc"
                },
                "start_index": {
                    "type": "integer",
                    "example": 0
                },
                "ticker": {
                    "type": "string",
                    "example": "AAPL"
                }
            }
        },
        "models.Comment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ticker_event_id": {
                    "description": "Check ticker and event to get proper comment seciton",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Event": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "event_type": {
                    "$ref": "#/definitions/models.EventType"
                },
                "event_types_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "Use timestamptz in Postgres",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.EventType": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Ticker": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "ticker": {
                    "type": "string"
                }
            }
        },
        "models.TickerEvent": {
            "type": "object",
            "properties": {
                "end_index": {
                    "type": "integer"
                },
                "event_id": {
                    "type": "string"
                },
                "events": {
                    "description": "optional: populated if joined",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Event"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "start_index": {
                    "type": "integer"
                },
                "ticker": {
                    "type": "string"
                }
            }
        },
        "models.UserInfo": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}